refactor: app 폴더 구성, App.tsx 이동
