+++ src/app/App.tsx
@@ -1,11 +1,12 @@
-import { BrowserRouter as Router } from "react-router-dom";
-import PostsManagerPage from "./pages/PostsManagerPage.tsx";
-import Footer from "./widgets/ui/Footer.tsx";
-import Header from "./widgets/ui/Header.tsx";
+import PostsManagerPage from "../pages/PostsManagerPage.tsx";
+import Footer from "../widgets/ui/Footer.tsx";
+import Header from "../widgets/ui/Header.tsx";
+import Providers from "./providers";
+import "./styles/global.css";
 
 const App = () => {
   return (
-    <Router>
+    <Providers>
       <div className="flex flex-col min-h-screen">
         <Header />
         <main className="flex-grow container mx-auto px-4 py-8">
@@ -13,7 +14,7 @@ const App = () => {
         </main>
         <Footer />
       </div>
-    </Router>
+    </Providers>
   );
 };
