+++ src/app/providers/index.tsx
@@ -1,0 +1,8 @@
+import { ReactNode } from "react";
+import { BrowserRouter as Router } from "react-router-dom";
+
+const Providers = ({ children }: { children: ReactNode }) => {
+  return <Router>{children}</Router>;
+};
+
+export default Providers;
